Making utilities extension methods
