+++ src/Diagnostics.ModelsAndUtils/Diagnostics.ModelsAndUtils.csproj
@@ -5,6 +5,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="DataTableUtility.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="System.Collections.Immutable" Version="1.3.1" />
   </ItemGroup>
